+++ src/Logger.ts
@@ -310,6 +310,20 @@ export class Logger extends EventEmitter {
         this.log(args, LogSeverity.DEPRECATE);
     }
 
+    public deprecateClass(deprecatedClass: string, alternative?: string): void {
+        let e: Error = new Error();
+
+        let args: any = [];
+        args.push(`Class ${deprecatedClass} is deprecated.`);
+        if (alternative) {
+            args.push(this._getDeprecatedAlternativeMessage(alternative));
+        }
+
+        args.push('\n\n');
+        args.push(e.stack);
+        this.log(args, LogSeverity.DEPRECATE);
+    }
+
     private _getDeprecatedMethodMessage(e: Error): string {
         let stack = e.stack .split('\n')[2].replace(/^\s+at\s+(.+?)\s.+/g, '$1');
         return `Method ${stack} is deprecated.`